requirements.txt
streamlit>=1.35.0
google-genai>=0.7.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
faiss-cpu>=1.8.0
pymupdf>=1.24.0
